+++ user_docs/en/changes.md
@@ -11,6 +11,9 @@ Windows 10 is the minimum Windows version supported.
 
 ### New Features
 
+* Added the possibility to report when multiple items can be selected in a list control.
+This can be enabled using the "Report when lists support multiple selection" setting in NVDA's object presentation settings. (#18365 @LeonarddeR)
+
 ### Changes
 
 * Added a button to the About dialog to copy the NVDA version number to the clipboard. (#18667)